Add rollback support to Manager, Snapshot and Plugins
Add new methodd Plugin.rollback_snapshot() and Snapshot.rollback()
and use these to implement Manager.rollback_snapshot_sets().

For the LVM2 plugin implementation this initiates a merge operation:
this takes immediate effect for inactive volumes and schedules the
merge to take place at the next volume activation for volumes that
are currently in use.