Propagate exceptions from Manager.create_snapshot_set()
The create_snapshot_set() method fails to propagate exceptions that are
raised during snapshot creation. This leeds to it incorrectly reporting
success when snapshot set creation has failed. Fix this by re-raising
the original exception after performing cleanup tasks.

This change currently breaks four test cases in ManagerTests that
attempt to create more than one snapshot set. This will be fixed in a
follow-up commit that increases the size of the loop device allocated
for the test VG.