snapm: fix %FREE space accounting
The current implementation of the PERCENT_FREE size policy gives
confusing and incorrect behavior when multiple snapshots are taken in
the same VG: for e.g. with 25%FREE the first snapshot receives 25% of
the initial free space, the second gets 25% of what's left after
allocating the first, and so on.

Introduce a notion of snapshot set create transactions in the Plugin
class and calculate all the snapshot sizes in the initial
Plugin.check_create_snapshot() pass.

This gives consistent behaviour where each snapshot receives the same
fraction of the initial available space.