snapm: add a size property to Snapshot
Add a property that reflects the size in bytes of a snapshot: this is
either the device size for snapshot providers that dynamically allocate
space to the snapshot or the backstore size for providers that have
fixed sized snapshots (i.e. lvm2-cow). For LVM2 snapshots this is the
value of the 'lv_size' report field.