snapm: add SnapmPluginError and improve error message on set create
When a plugin operation fails with e.g. SnapmCalloutError, catch
the initial exception and then re-raise a new exception type,
SnapmPluginError with a more descriptive error message.

This changes failure in lvcreate from:

ERROR - Error creating snapshot set member backup-plus: lvcreate failed with:   Size is not a multiple of 512. Try using 2384881152 or 2384881664.
  Invalid argument for --size: 2384881255b
  Error during parsing of command line.

ERROR - Command failed: lvcreate failed with:   Size is not a multiple of 512. Try using 2384881152 or 2384881664.
  Invalid argument for --size: 2384881255b
  Error during parsing of command line.

Which just duplicates the LVM2 error message to:

ERROR - Error creating snapshot set member backup-plus: lvcreate failed with:   Size is not a multiple of 512. Try using 2384881152 or 2384881664.
  Invalid argument for --size: 2384881255b
  Error during parsing of command line.

ERROR - Command failed: Could not create all snapshots for set backup-plus