snapm: generalise revert check for mounted fs to snapshot and origin
The snapshot is considered "in use" for revert purposes if either the
snapshot or the origin is currently mounted and in use. Add
snapshot_mounted properties to Snapshot and SnapshotSet and a new
mounted property which is True when either the snapshot or its origin is
in use.

Test SnapshotSet.mounted in snapm.manager and snapm.command when
determining whether a to-be-reverted snapshot set is in use or not.

Resolves: #61