manager: handle errors when initialising plugins
Allow plugins to be disabled when their dependencies are not met:
a plugin can raise SnapmNotFoundError in __init__() to indicate that the
required dependencies are missing. This is not treated as an error. If
dependencies are present but for some reason unusable (for e.g. minimum
version requirements are not met) plugins should raise SnapmPluginError
and this will be reported as an error.