lvm2: refactor helper functions as Lvm2 plugin methods
Since lvm2 on Ubuntu-24.04 does not support the --reportformat json_std
argument refactor the lvm2 plugin helper functions that call out to lvm2
reporting commands to be methods of the Lvm2 plugin class.

This allows the plugin to detect the lvm2 version when it is initialised
and set a private property, _json_fmt to the required --reportformat
string.

Subsequent calls to the helpers can then use this value to issue the
appropriate command line arguments for the running lvm2 version.

Amend test cases for the corresponding functions to use the appropriate
object type for helper unit tests.

Related: #110