lvm2: filter out LVM2 environment variables when calling lvm commands
There are numerous LVM2 environment variables (see lvm(8)) that can
influence the behaviour of the commands that snapm relies on in the lvm2
plugins.

Filter the environment when initialising the plugin to discard any
unwanted variables: the LVM_SYSTEM_DIR environment variable is
explicitly excluded since it is used in the test suite and can be used
legitimately to direct snapm to an alternate LVM configuration.

Resolves: #112