tests: refactor CalendarSpec tests to use unittest.subTest()
There are two test cases in CalendarSpecTests that each have a large
(~50 entries) table of individual calls to either the _test_one() or
_test_next() methods.

Refactor this so that instead of individual calls we iterate over a list
of tuples comprising the input data and expected output, and call the
test methods under the context manager 'with self.subTest()'.

This generates a unique index for each subtest which is reported with
any failures, as well as a custom message that includes the input data
to the test.

This makes debugging easier since it makes it explicitly clear which
test case has failed.

Related: #122