snapm: support SnapshotSet names with basename and index
Extend the naming scheme for snapshot sets to allow snapshot sets with a
common basename and an index that is incremented for each new instance
of the basename.

Indexed snapshot sets are identified by a trailing numeric value
separated from the basename with a dot:

   hourly.0 -> name=hourly.0 basename=hourly index=0
   hourly.1 -> name=hourly.1 basename=hourly index=1
   hourly.2 -> name=hourly.2 basename=hourly index=2
   hourly.3 -> name=hourly.3 basename=hourly index=3
   ...

The values are parsed from the name when initialising Snapshot and
SnapshotSet objects and are exposed as the read-only properties
'basename' and 'index'.

Resolves: #158