snapm: add autoindex support to Manager.create_snapshot_set()
Support recurring snapshot sets with a basename and index: add a new
'autoindex' kwarg to the create_snapshot_set() method. If `True`, an
index will be generated and appended to the `name`, separated by a
single '.' character.

For e.g.:

  name="hourly" with existing sets "hourly.0", "hourly.1", "hourly.2",
  the next index is 3.

The index is generated by finding all existing snapshot sets that match
the given basename, and returning one plus the greatest index found:
indexes are generated sequentially but may be deleted in any order,
depending on the user and any configured garbage collection policy.

Resolves: #159