snapm: refactor plugin loader and move to snapm.manager._loader
Simplify the plugin loading logic and move it into a separate module:
the load_plugins() function now returns a list of classes that can be
instantiated from the plugin package.

This removes the need for the ImporterHelper and PluginRegistry
metaclass and allows the Plugin class to be moved into the
snapm.manager.plugins package.

Resolves: #192