dist: update systemd timer templates to always include OnCalendar
Not setting a valid 'On*' option to define a timer trigger condition
leads to unwanted behaviour when attempting to operate on non-existent
timer unit instances:

  # systemctl status snapm-gc@foobar.timer
  ○ snapm-gc@foobar.timer - Scheduled foobar snapshot set cleanup
       Loaded: bad-setting (Reason: Unit snapm-gc@foobar.timer has a bad unit file setting.)
       Active: inactive (dead)
      Trigger: n/a
     Triggers: ● snapm-gc@foobar.service

  May 21 13:12:13 localhost.localdomain systemd[1]: snapm-gc@foobar.timer: Timer unit lacks value setting. Refusing.

This is somewhat misleading: snamp-gc@foobar.timer *does not exist*:
there is no instance symlink of that name but systemd noticies that the
template lacks a defined On* value and complains.

This may be confusing to users or trigger alerts based on error logs:
define a dumby "daily" value in the base template unit that is overriden
on a per-instance basis via drop-in files.

With this change the errors are no longer triggered, and the systemctl
status output looks like:

  # systemctl status snapm-create@bazquux.timer || echo $?
  ○ snapm-create@bazquux.timer - Scheduled bazquux snapshot set creation
       Loaded: loaded (/usr/lib/systemd/system/snapm-create@.timer; disabled; preset: disabled)
       Active: inactive (dead)
      Trigger: n/a
     Triggers: ● snapm-create@bazquux.service
  3

  (exit status 3 means "unit is not active")

It would be ideal if systemd simply recognised the instance as
non-existent and responded like it would for any other non-existent unit
definition:

  # systemctl status bazquux.timer || echo $?
  Unit bazquux.timer could not be found.
  4

  (exit status 4 means "no such unit - systemctl(1))

Some similar issues have been raised in the past but are still open as
of today:

  https://github.com/systemd/systemd/issues/25708
  https://github.com/systemd/systemd/issues/25680