progress: limit max FPS in Progress like Throbber
Eagerly updating the progress bar on every Progress.progress() call
hammers the terminal and causes flicker when working with large total:
use the same mechanism as Throbber to limit the FPS to something
reasonable.

The compute_diffs() progress (which is normally very quick but takes ~30s
on bigger trees with 600k total paths and 400k paths in the tree dicts)
is flickery as heck. Here's how we can support it without the problems I
noted earlier in the issue:

  * Save the last status line in the class, even if we decide not to
    output it due to FPS limiting

  * On cancel (e.g. KeyboardInterrupt) re-draw the last status - that
    way the user knows what path we got stuck on if something hangs.

Resolves: #685

Signed-off-by: Bryn M. Reeves <bmr@redhat.com>