fsdiff: use BinaryContentDiffer as the differ-of-last-restort
If we don't have anyFileTypeInfo for an entry the current
old_entry/new_entry we fall back to TextContextDiffer:

611         try:
612             differ = (
613                 self.get_differ_for_file(file_type_info)
614                 if file_type_info is not None
615                 else TextContentDiffer()
616             )
617             return differ.generate_diff(old_path, new_path, old_entry, new_entry)

This is the root cause of the memory pressure problems on low-memory
systems that prompted #786, #789, #790, #798, and #800. These are all
good-to-haves but the reason we were generating 3GiB RSS on these 4GiB
systems was that without -f / --file-type we attempted to generate text
content diffs for everything that changed - hundreds of xz compressed
kernel modules for e.g. (since an add counts as modified and we diff
against /dev/null):

Saving cache:  38% [==========------------------] (Saving record 17525)
/usr/lib/modules/6.17.12-300.fc43.x86_64/kernel/drivers/infiniband/core/ib_core.ko.xz
    diff_type: unified
      old_content: ''
      new_content: '�7zXZ\x00\x00\x01i"�6\x04�Ч\x0f...'
      diff_data: <1923 items>
      summary: 0 deletions, 1923 additions
      has_changes: True
      error_message:
Saving cache:  38% [==========------------------] (Saving record 17526)
/usr/lib/modules/6.17.12-300.fc43.x86_64/kernel/drivers/infiniband/core/ib_umad.ko.xz
    diff_type: unified
      old_content: ''
      new_content: '�7zXZ\x00\x00\x01i"�6\x04���...'
      diff_data: <349 items>
      summary: 0 deletions, 349 additions
      has_changes: True
      error_message:
Saving cache:  38% [==========------------------] (Saving record 17527)
/usr/lib/modules/6.17.12-300.fc43.x86_64/kernel/drivers/infiniband/core/ib_uverbs.ko.xz
    diff_type: unified
      old_content: ''
      new_content: '�7zXZ\x00\x00\x01i"�6\x04���...'
      diff_data: <948 items>
      summary: 0 deletions, 948 additions
      has_changes: True
      error_message:

Since we now have magic-less best effort type guessing (#801 / #803)
this should be unreachable dead code, but for belts-and-braces let's
switch to the BinaryContentDiffer as the fallback default: it is safer
since it never attempts to compute actual diffs.

Resolves: #819

Signed-off-by: Bryn M. Reeves <bmr@redhat.com>